Prepare codebase for release with educational enhancements
- Renamed module from nbx_agent_compliance to netbox_agent_compliance
- Updated pyproject.toml package name and entry points
- Updated README.md with new command names
- Added SECURITY.md and LICENSE files
- Added comprehensive educational comments throughout codebase to help
  developers understand the agent pattern, MCP integration, and NetBox
  compliance checking
- Applied ruff formatting for consistency

These changes prepare the repository for public release as a learning
resource for building AI agents with NetBox MCP server.
