+++ SECURITY.md
@@ -1,0 +1,5 @@
+# Security Policy
+
+## Reporting a Vulnerability
+
+Please send any suspected vulnerability report to security@netboxlabs.com
